+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+### 0.3.2
+
+* Added help to console command
+* Fixed notice thrown by boost versions
+
 ### 0.3.1
 
 * Fixed issue with detached HEAD where pull would not be properly merged